[Fix] 케이크점 생성 API 수정

+++ server/src/main/java/cakeit/server/entity/CakeStoreEntity.java
@@ -25,7 +25,7 @@ public class CakeStoreEntity extends AbstractEntity {
 	private String storeName;
 
    	@Column(name = "STORE_SCORE")
-	private Long storeScore;
+	private Double storeScore;
 
    	@Column(name = "STORE_TIME")
 	private String storeTime;
